requirements.txt
python-telegram-bot==13.4.1
requests==2.25.1
python-Levenshtein==0.12.2